パッケージ構成を変更
