+++ app/src/main/java/com/example/androidgithubsearch/data/api/GitHubApiService.kt
@@ -1,4 +1,4 @@
-package com.example.androidgithubsearch.api
+package com.example.androidgithubsearch.data.api
 
 import retrofit2.http.GET
 import retrofit2.http.Path
