+++ app/src/main/java/com/example/androidgithubsearch/data/database/entity/FavoriteRepositoryEntity.kt
@@ -1,4 +1,4 @@
-package com.example.androidgithubsearch.database.entity
+package com.example.androidgithubsearch.data.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
