+++ app/src/main/java/com/example/androidgithubsearch/data/database/entity/UserRepositoryEntity.kt
@@ -1,8 +1,8 @@
-package com.example.androidgithubsearch.database.entity
+package com.example.androidgithubsearch.data.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.example.androidgithubsearch.ui.adapter.UserRepositoryItem
+import com.example.androidgithubsearch.ui.fragment.userrepository.UserRepositoryItem
 import java.util.Date
 
 @Entity
