+++ app/src/main/java/com/example/androidgithubsearch/data/sharedpreferences/SharedPreferencesKeys.kt
@@ -1,4 +1,4 @@
-package com.example.androidgithubsearch.util
+package com.example.androidgithubsearch.data.sharedpreferences
 
 enum class SharedPreferencesKeys(val value: String) {
     USER_NAME("username")
