+++ app/src/main/java/com/example/androidgithubsearch/di/HiltModule.kt
@@ -2,8 +2,8 @@ package com.example.androidgithubsearch.di
 
 import android.content.Context
 import androidx.room.Room
-import com.example.androidgithubsearch.api.GitHubApiService
-import com.example.androidgithubsearch.database.AppDatabase
+import com.example.androidgithubsearch.data.api.GitHubApiService
+import com.example.androidgithubsearch.data.database.AppDatabase
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
