+++ app/src/main/java/com/example/androidgithubsearch/ui/fragment/searchrepository/SearchRepositoryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.androidgithubsearch.ui.fragment
+package com.example.androidgithubsearch.ui.fragment.searchrepository
 
 import android.content.Context
 import android.os.Bundle
@@ -12,8 +12,6 @@ import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.DividerItemDecoration
 import com.example.androidgithubsearch.databinding.FragmentSearchRepositoryBinding
 import com.example.androidgithubsearch.ui.activity.WebViewActivity
-import com.example.androidgithubsearch.ui.adapter.SearchRepositoryAdapter
-import com.example.androidgithubsearch.ui.viewmodel.SearchRepositoryFragmentViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
