+++ app/src/main/java/com/example/androidgithubsearch/ui/fragment/searchrepository/SearchRepositoryFragmentViewModel.kt
@@ -1,12 +1,11 @@
-package com.example.androidgithubsearch.ui.viewmodel
+package com.example.androidgithubsearch.ui.fragment.searchrepository
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.androidgithubsearch.repository.GitHubRepository
-import com.example.androidgithubsearch.ui.adapter.SearchRepositoryItem
+import com.example.androidgithubsearch.data.repository.GitHubRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
