+++ app/src/main/java/com/example/androidgithubsearch/ui/fragment/userrepository/UserRepositoryFragmentViewModel.kt
@@ -1,15 +1,14 @@
-package com.example.androidgithubsearch.ui.viewmodel
+package com.example.androidgithubsearch.ui.fragment.userrepository
 
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.androidgithubsearch.database.entity.UserRepositoryEntity
-import com.example.androidgithubsearch.repository.GitHubRepository
-import com.example.androidgithubsearch.ui.adapter.UserRepositoryItem
-import com.example.androidgithubsearch.util.SharedPreferencesKeys
-import com.example.androidgithubsearch.util.SharedPreferencesUtil
+import com.example.androidgithubsearch.data.database.entity.UserRepositoryEntity
+import com.example.androidgithubsearch.data.repository.GitHubRepository
+import com.example.androidgithubsearch.data.sharedpreferences.SharedPreferencesKeys
+import com.example.androidgithubsearch.data.sharedpreferences.SharedPreferencesUtil
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
