+++ app/src/main/res/layout/user_repository_row_item.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="repositoryItem"
-            type="com.example.androidgithubsearch.ui.adapter.UserRepositoryItem" />
+            type="com.example.androidgithubsearch.ui.fragment.userrepository.UserRepositoryItem" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
